chore: move unit related state methods in app domain
Following in the lines of https://github.com/juju/juju/pull/19049 this
patch moves the unit-related methods in the state layer of the
application domain to its own new file unit.go.
Tests have also been moved to unit_test.go.
